requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap4==2024.10
Django==5.2.2
django-crispy-forms==2.4
django-taggit==6.1.0
pillow==11.2.1
reportlab==4.4.1
sqlparse==0.5.3
Unidecode==1.4.0
